refactor: move ui helpers to module
