+++ test/sanity.test.js
@@ -1,3 +1,4 @@
+/* global describe, it */
 import { expect } from 'chai';
 
 describe('sanity', () => {
